backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.2.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.12.0
google-cloud-build==3.20.0
google-generativeai==0.7.2
Jinja2==3.1.6
python-dotenv==1.0.1
gunicorn==22.0.0
google-cloud-run==0.12.0
